# openapi.yaml
openapi: 3.1.0
info:
  title: Sistema IoT Riego Agrícola
  version: 0.1.0
paths:
...
  /api/v1/clients/me/notification-settings:
    get:
      tags:
      - Clients
      summary: Get My Notification Settings
      operationId: get_my_notification_settings_api_v1_clients_me_notification_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientNotificationSettingsResponse'
      security:
      - OAuth2PasswordBearer: []
    patch:
      tags:
      - Clients
      summary: Update My Notification Settings
      operationId: update_my_notification_settings_api_v1_clients_me_notification_settings_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientNotificationSettingsUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientNotificationSettingsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/clients/me/dashboard-preferences:
    get:
      tags:
      - Clients
      summary: Get My Dashboard Preferences
      operationId: get_my_dashboard_preferences_api_v1_clients_me_dashboard_preferences_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardPreferencesResponse'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/clients:
    get:
      tags:
      - Clients
      summary: List Clients
      operationId: list_clients_api_v1_clients_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ClientResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Clients
      summary: Create Client
      operationId: create_client_api_v1_clients_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/clients/{client_id}:
    get:
      tags:
      - Clients
      summary: Get Client
      operationId: get_client_api_v1_clients__client_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Clients
      summary: Update Client
      operationId: update_client_api_v1_clients__client_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Clients
      summary: Delete Client
      operationId: delete_client_api_v1_clients__client_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/clients/{client_id}/dashboard-preferences:
    get:
      tags:
      - Clients
      summary: Get Dashboard Preferences
      operationId: get_dashboard_preferences_api_v1_clients__client_id__dashboard_preferences_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardPreferencesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Clients
      summary: Update Dashboard Preferences
      operationId: update_dashboard_preferences_api_v1_clients__client_id__dashboard_preferences_put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardPreferencesUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardPreferencesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ClientCreate:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        full_name:
          type: string
          title: Full Name
        company_name:
          type: string
          title: Company Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      required:
      - email
      - password
      - full_name
      - company_name
      title: ClientCreate
    ClientNotificationSettingsResponse:
      properties:
        notifications_enabled:
          type: boolean
          title: Notifications Enabled
      type: object
      required:
      - notifications_enabled
      title: ClientNotificationSettingsResponse
    ClientNotificationSettingsUpdate:
      properties:
        notifications_enabled:
          type: boolean
          title: Notifications Enabled
      type: object
      required:
      - notifications_enabled
      title: ClientNotificationSettingsUpdate
    ClientResponse:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        company_name:
          type: string
          title: Company Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        notifications_enabled:
          type: boolean
          title: Notifications Enabled
        user:
          anyOf:
          - $ref: '#/components/schemas/UserResponse'
          - type: 'null'
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - company_name
      - notifications_enabled
      - created_at
      - updated_at
      title: ClientResponse
    ClientUpdate:
      properties:
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        company_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
      type: object
      title: ClientUpdate
...
    DashboardPreferencesResponse:
      properties:
        client_id:
          type: integer
          title: Client Id
        cards:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Cards
      type: object
      required:
      - client_id
      - cards
      title: DashboardPreferencesResponse
    DashboardPreferencesUpdate:
      properties:
        cards:
          items:
            type: string
          type: array
          title: Cards
      additionalProperties: false
      type: object
      required:
      - cards
      title: DashboardPreferencesUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_ClientResponse_:
      properties:
        page:
          type: integer
          title: Page
        per_page:
          type: integer
          title: Per Page
        total:
          type: integer
          title: Total
        data:
          items:
            $ref: '#/components/schemas/ClientResponse'
          type: array
          title: Data
      type: object
      required:
      - page
      - per_page
      - total
      - data
      title: PaginatedResponse[ClientResponse]
...
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          title: Email
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - full_name
      - role
      - is_active
      - created_at
      - updated_at
      title: UserResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/docs-login